Remove recordLanguages ...
It really does not make sense to have "recordLanguages" at this level since it is not likely that the collection object would be available is one set of languages and the record objects would be available in a completely different set of languages.  This record-level language information is better restricted to the record itself which is it in recordGeoJSON.json.